backend/requirements.txt
mysql-connector-python==8.0.31
python-dotenv==1.0.0
Flask==2.3.3
Flask-CORS==4.0.0